Update misc_arcpy_ops.py
Fixed a glitch with enforceLabels function.

+++ Scripts/misc_arcpy_ops.py
@@ -454,14 +454,14 @@ def enforceLabels(feature_item : str) -> None:
                             update_row = True
                     elif not row[0] is None:
                         try:
-                            if (new_str := str(float(row[0]))) != row[2]:
+                            if (new_str := str(int(row[0]))) != row[2]:
                                 row[2] = new_str
                                 update_row = True
                         except ValueError:
                             pass
                     elif not row[2] is None:
                         try:
-                            row[0] = float(row[2])
+                            row[0] = int(row[2])
                             update_row = True
                         except ValueError:
                             pass
